+++ deal-api-owner/src/main/kotlin/com/chaltteok/owner/dailystock/enums/DailyStockStatusType.kt
@@ -1,0 +1,4 @@
+package com.chaltteok.owner.enums
+
+enum class DailyStockStatusType {
+}
