+++ deal-api-owner/src/main/kotlin/com/chaltteok/owner/dailystock/enums/DailyStockType.kt
@@ -1,0 +1,4 @@
+package com.chaltteok.owner.dailystock.enums
+
+class DailyStockType {
+}
