+++ deal-api-owner/src/main/kotlin/com/chaltteok/owner/product/enums/StockType.kt
@@ -1,0 +1,6 @@
+package com.chaltteok.owner.product.enums
+
+enum class StockType(name: String) {
+    EVENT("EVENT"),
+    NORMAL("NORMAL")
+}
