+++ deal-api-owner/src/main/kotlin/com/chaltteok/owner/product/service/ProductService.kt
@@ -1,0 +1,8 @@
+package com.chaltteok.owner.product.service
+
+import com.chaltteok.owner.product.dto.ProductRegisterRequest
+import org.springframework.web.multipart.MultipartFile
+
+interface ProductService {
+    fun registerProduct(productRegisterRequest: ProductRegisterRequest, image: MultipartFile?)
+}
