--- deal-api-owner/src/main/kotlin/com/chaltteok/owner/service/ProductService.kt
@@ -1,8 +1,0 @@
-package com.chaltteok.owner.service
-
-import com.chaltteok.owner.dto.ProductRegisterRequest
-import org.springframework.web.multipart.MultipartFile
-
-interface ProductService {
-    fun registerProduct(productRegisterRequest: ProductRegisterRequest,image: MultipartFile?)
-}
